Refactor(RequestHandler) : MimeType 도출 로직의 메서드화
MimeType 도출 테스트를 위해 해당 로직을 메서드로 만들었다.
